requirements.txt
ExifRead==2.3.2
gpsphoto==2.2.3
numpy==1.19.5
piexif==1.1.3
Pillow==8.2.0
